Client: split up components (refactoring)
No changes in functionality, just restructuring
