+++ client/src/main/scala/io/github/mahh/doko/client/components/StringComponents.scala
@@ -1,0 +1,31 @@
+package io.github.mahh.doko.client.components
+
+import com.raquo.laminar.api.L.*
+import io.github.mahh.doko.shared.msg.MessageToServer.SetUserName
+object StringComponents {
+
+  def nameInput(
+    disabledSignal: SignalSource[Boolean],
+    commandSink: SetUserName => Unit
+  ): Div =
+    div(
+      label(
+        "Your name: ",
+        hidden <-- disabledSignal
+      ),
+      input(
+        onMountFocus,
+        placeholder := "Enter your name here",
+        hidden <-- disabledSignal,
+        onInput.mapToValue.map(SetUserName.apply) --> commandSink
+      ),
+      p()
+    )
+
+  def announcement(
+    contentObservable: Observable[Option[String]]
+  ): Div =
+    div(
+      child.maybe <-- contentObservable.map(_.map(txt => span(txt)))
+    )
+}
